Ajout boutons et trajectoires

+++ include/inregister.h
@@ -6,12 +6,12 @@ class inregister{
     void Setup(int latchPin, int clockPin, int dataPin);
     bool readInput(int bitNumber);
     void fetch();
-    unsigned int readValue();
+    unsigned long readValue();
   private:
     int latchPin;
     int clockPin;
     int dataPin;
-    unsigned int result;
+    unsigned long result;
 };
 
 #endif
